Refactor viz summaries and stabilize validation checks
