+++ docs/sg_execution_times.rst
@@ -6,7 +6,7 @@
 
 Computation times
 =================
-**00:12.590** total execution time for 17 files **from all galleries**:
+**00:04.874** total execution time for 17 files **from all galleries**:
 
 .. container::
 
@@ -32,54 +32,54 @@ Computation times
    * - Example
      - Time
      - Mem (MB)
-   * - :ref:`sphx_glr_auto_examples_dss_plot_01_dss_fundamentals.py` (``../examples/dss/plot_01_dss_fundamentals.py``)
-     - 00:06.411
+   * - :ref:`sphx_glr_auto_examples_dss_plot_07_spectrogram_dss.py` (``..\examples\dss\plot_07_spectrogram_dss.py``)
+     - 00:04.874
      - 0.0
-   * - :ref:`sphx_glr_auto_examples_dss_plot_04_spectral_dss.py` (``../examples/dss/plot_04_spectral_dss.py``)
-     - 00:06.179
-     - 0.0
-   * - :ref:`sphx_glr_auto_examples_dss_plot_02_artifact_correction.py` (``../examples/dss/plot_02_artifact_correction.py``)
+   * - :ref:`sphx_glr_auto_examples_dss_plot_01_dss_fundamentals.py` (``..\examples\dss\plot_01_dss_fundamentals.py``)
      - 00:00.000
      - 0.0
-   * - :ref:`sphx_glr_auto_examples_dss_plot_03_evoked_responses.py` (``../examples/dss/plot_03_evoked_responses.py``)
+   * - :ref:`sphx_glr_auto_examples_dss_plot_02_artifact_correction.py` (``..\examples\dss\plot_02_artifact_correction.py``)
      - 00:00.000
      - 0.0
-   * - :ref:`sphx_glr_auto_examples_dss_plot_05_periodic_dss.py` (``../examples/dss/plot_05_periodic_dss.py``)
+   * - :ref:`sphx_glr_auto_examples_dss_plot_03_evoked_responses.py` (``..\examples\dss\plot_03_evoked_responses.py``)
      - 00:00.000
      - 0.0
-   * - :ref:`sphx_glr_auto_examples_dss_plot_06_temporal_dss.py` (``../examples/dss/plot_06_temporal_dss.py``)
+   * - :ref:`sphx_glr_auto_examples_dss_plot_04_spectral_dss.py` (``..\examples\dss\plot_04_spectral_dss.py``)
      - 00:00.000
      - 0.0
-   * - :ref:`sphx_glr_auto_examples_dss_plot_07_spectrogram_dss.py` (``../examples/dss/plot_07_spectrogram_dss.py``)
+   * - :ref:`sphx_glr_auto_examples_dss_plot_05_periodic_dss.py` (``..\examples\dss\plot_05_periodic_dss.py``)
      - 00:00.000
      - 0.0
-   * - :ref:`sphx_glr_auto_examples_dss_plot_08_blind_source_separation.py` (``../examples/dss/plot_08_blind_source_separation.py``)
+   * - :ref:`sphx_glr_auto_examples_dss_plot_06_temporal_dss.py` (``..\examples\dss\plot_06_temporal_dss.py``)
      - 00:00.000
      - 0.0
-   * - :ref:`sphx_glr_auto_examples_dss_plot_09_custom_bias.py` (``../examples/dss/plot_09_custom_bias.py``)
+   * - :ref:`sphx_glr_auto_examples_dss_plot_08_blind_source_separation.py` (``..\examples\dss\plot_08_blind_source_separation.py``)
      - 00:00.000
      - 0.0
-   * - :ref:`sphx_glr_auto_examples_dss_plot_10_benchmarking.py` (``../examples/dss/plot_10_benchmarking.py``)
+   * - :ref:`sphx_glr_auto_examples_dss_plot_09_custom_bias.py` (``..\examples\dss\plot_09_custom_bias.py``)
      - 00:00.000
      - 0.0
-   * - :ref:`sphx_glr_auto_examples_dss_plot_11_wiener_masking.py` (``../examples/dss/plot_11_wiener_masking.py``)
+   * - :ref:`sphx_glr_auto_examples_dss_plot_10_benchmarking.py` (``..\examples\dss\plot_10_benchmarking.py``)
      - 00:00.000
      - 0.0
-   * - :ref:`sphx_glr_auto_examples_dss_plot_12_joint_dss.py` (``../examples/dss/plot_12_joint_dss.py``)
+   * - :ref:`sphx_glr_auto_examples_dss_plot_11_wiener_masking.py` (``..\examples\dss\plot_11_wiener_masking.py``)
      - 00:00.000
      - 0.0
-   * - :ref:`sphx_glr_auto_examples_zapline_plot_01_basic_usage.py` (``../examples/zapline/plot_01_basic_usage.py``)
+   * - :ref:`sphx_glr_auto_examples_dss_plot_12_joint_dss.py` (``..\examples\dss\plot_12_joint_dss.py``)
      - 00:00.000
      - 0.0
-   * - :ref:`sphx_glr_auto_examples_zapline_plot_02_parameter_tuning.py` (``../examples/zapline/plot_02_parameter_tuning.py``)
+   * - :ref:`sphx_glr_auto_examples_zapline_plot_01_basic_usage.py` (``..\examples\zapline\plot_01_basic_usage.py``)
      - 00:00.000
      - 0.0
-   * - :ref:`sphx_glr_auto_examples_zapline_plot_03_epoched_data.py` (``../examples/zapline/plot_03_epoched_data.py``)
+   * - :ref:`sphx_glr_auto_examples_zapline_plot_02_parameter_tuning.py` (``..\examples\zapline\plot_02_parameter_tuning.py``)
      - 00:00.000
      - 0.0
-   * - :ref:`sphx_glr_auto_examples_zapline_plot_04_adaptive_mode.py` (``../examples/zapline/plot_04_adaptive_mode.py``)
+   * - :ref:`sphx_glr_auto_examples_zapline_plot_03_epoched_data.py` (``..\examples\zapline\plot_03_epoched_data.py``)
      - 00:00.000
      - 0.0
-   * - :ref:`sphx_glr_auto_examples_zapline_plot_05_adaptive_advanced.py` (``../examples/zapline/plot_05_adaptive_advanced.py``)
+   * - :ref:`sphx_glr_auto_examples_zapline_plot_04_adaptive_mode.py` (``..\examples\zapline\plot_04_adaptive_mode.py``)
+     - 00:00.000
+     - 0.0
+   * - :ref:`sphx_glr_auto_examples_zapline_plot_05_adaptive_advanced.py` (``..\examples\zapline\plot_05_adaptive_advanced.py``)
      - 00:00.000
      - 0.0
